+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Change Log
 
 
+## [v1.4.5](https://plugins.jetbrains.com/plugin/10907-mvpautocodeplus)(2020-10-17)
+- Adapter Android studio 4.1,fix create layout file error.
+
 ## [v1.4.4](https://plugins.jetbrains.com/plugin/10907-mvpautocodeplus)(2020-9-27)
 - Fix Bug: Import R File package name incorrect
 
 ## [v1.4.3](https://plugins.jetbrains.com/plugin/10907-mvpautocodeplus)(2020-6-4)
-- Adapte Android studio 4.0
+- Adapter Android studio 4.0
 
 ## [v1.4.2](https://plugins.jetbrains.com/plugin/10907-mvpautocodeplus)(2020-4-25)
 - Add layout file create.